refactor plugin structure, update gradle
