+++ buildSrc/src/main/kotlin/plugin/GenerateOpenGLTask.kt
@@ -1,0 +1,17 @@
+package plugin
+
+import opengl.OpenGLGenerator
+import org.gradle.api.tasks.GradleBuild
+import org.gradle.api.tasks.OutputDirectory
+import org.gradle.api.tasks.TaskAction
+
+open class GenerateOpenGLTask : GradleBuild() {
+    @OutputDirectory
+    val outputDir = project.buildDir.resolve("generated-src")
+
+    @TaskAction
+    fun exec() {
+        OpenGLGenerator.generate(outputDir)
+    }
+}
+
